# resources/swagger.yaml
openapi: 3.0.0
info:
  title: API de Treinos de Musculação
  version: 1.0.0
  description: API para criação e consulta de treinos de musculação para diferentes faixas etárias e IMC.
servers:
  - url: http://localhost:3000
paths:
  /workout:
    post:
      summary: Criação de treino personalizado
      description: Calcula o IMC do aluno e retorna as séries e carga conforme regras de idade e IMC.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                age:
                  type: integer
                weight:
                  type: number
                height:
                  type: number
      responses:
        '200':
          description: Treino criado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  imc:
                    type: number
                  mensagem:
                    type: string
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /:
    get:
      summary: Endpoint de teste
      responses:
        '200':
          description: API rodando
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '500':
          description: Erro interno
  /auth/register:
    post:
      summary: Registrar usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: Usuário registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Erro de validação
  /auth/login:
    post:
      summary: Login do usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login realizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '401':
          description: Credenciais inválidas
  /auth/profile:
    get:
      summary: Retorna perfil do usuário autenticado
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Perfil do usuário
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Não autenticado
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [student, trainer]
        age:
          type: integer
        weight:
          type: number
        height:
          type: number
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT